Fix bug in DF equality comparison and simplify implementation
There's currently a bug in `assertDataFrameDataEquals` (and hence
`assertDataFrameNoOrderEquals`) because it doesn't use Spark's null-safe
equality comparison. Specifically, the `.filter` in:

```
  val diff = expectedElementsCount
    .join(resultElementsCount, expected.columns, "full_outer")
    .filter(col(expectedCol) =!= col(actualCol))
```

will only keep rows where both `expectedCol` and `actualCol` are
non-null. Thus, if a row is present once in the `expected` DF, and not present in
the `result`, the `actualCol` will be `null` and this will cause the
column to be removed and not be reported as a difference.

The test I've added fails with the current implementation.

I've also replaced the implementation with a shorter version that uses
Spark 2.4's `intersectAll` directly. I think this version is slightly
simpler to follow because it uses set operations directly.

I'm not sure how the versioning works - I see that there are different
folders in `core/src/main`. Should I put the new implementation in a
`2.4` folder since it uses 2.4-specific functions? If so, do we also
need to backport a fix for older versions as well?

cc @holdenk